==> discovery/06_singlecell_TME.Rmd <==
---
title: "06_singlecell_tumor"
author: "Lasse Meyer"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Software requirements

```{r lib, message=FALSE,warning=FALSE}
#Data containers
library(SpatialExperiment)
library(SingleCellExperiment)

#Data processing
library(scuttle)
library(scater)
library(scran)
library(imcRtools)

#Data visualization
library(cytomapper)
library(dittoSeq)
library(pheatmap)
library(cowplot)
library(ComplexHeatmap)
library(circlize)
library(wesanderson)

#Utilities
library(tidyverse)
library(openxlsx)
library(stringr)
library(patchwork)
library(parallel)
library(BiocParallel)
library(BiocNeighbors)
library(BiocSingular)
library(tiff)
library(ggrepel)
library(EBImage)
library(viridis)
library(mclust)
library(Rphenograph)
library(igraph)
library(RColorBrewer)
library(foreach)
library(clValid)
library(bluster)
library(psych)
library(corrplot)
library(psych)

#Batch correction
library(batchelor)
```

# Single-cell analysis - Pipeline 3 - TNBC TME phenotypes

Here, we will perform the third single cell analysis steps, including phenotyping of TME cells.

## Read in data

First, we will read in the `SingleCellExperiment` object containing the cleaned, spillover- and (batch-corrected) single-cell data. 

```{r read-data-batch-correction, message=FALSE}
sce <- readRDS("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/04_sce_TNBC_Clus_com.rds")
```

## Filter sce for tumor cells
Here, we will filter for TME cells based on the cluster category and GMM approach.

```{r filter sce, message=FALSE}
#based on cluster categories (n=525537)
sce <- sce[,colData(sce)$analysis_cat == "non_tumor"]
sce
```

# 1. Clustering 

## Define markers for clustering

```{r tumor_markers, message = FALSE, warning = FALSE}
cluster_channel = c("SMA","CD68","CD3","CD20","CD11c","CD20","CD8a","CD4","CD31_vWF","CD15", "HLADR")

other_epi_channel = c("pH3","H3K27me3","Bcl2", "c_Cas3_PARP","S6", "CD44","CAIX","Ki67", "Vimentin","Fibronectin","panCK","Ecad","CK5","CK8_18","KRT14","cMYC","HER2","ER","PR","p53","AR","EGFR","p_mTOR","CK7","GATA3","CD45")

#Add to sce
rowData(sce)$cluster_channel <- (rowData(sce)$clean_target %in% cluster_channel)
rowData(sce)$other_epi_channel <- (rowData(sce)$clean_target %in% other_epi_channel)
```

## Run unsupervised clustering
Next, we will run Rphenograph (Jaccard-based weights, Louvain clustering as community detection algorithm) clustering with different resolution parameters (k=30,50,75). 

```{r overcluster, message=FALSE}
library(Rphenoannoy)
set.seed(22)

mat <- t(assay(sce, "exprs")[rowData(sce)$cluster_channel,])

#k=30
out <- Rphenoannoy(mat, k = 30)

clusters <- factor(membership(out[[2]]))

sce$cluster_30 <- clusters

unique(sce$cluster_30)

#Save SCEs after clustering
saveRDS(sce, "/mnt/central_nas/tnbc_volume/TNBC/data_analysis/06_sce_TNBC_TME_com.rds")

gc()

#k=50
out <- Rphenoannoy(mat, k = 50)

clusters <- factor(membership(out[[2]]))

sce$cluster_50 <- clusters

unique(sce$cluster_50)

#Save SCEs after clustering
saveRDS(sce, "/mnt/central_nas/tnbc_volume/TNBC/data_analysis/06_sce_TNBC_TME_com.rds")

gc()

#k=70
out <- Rphenoannoy(mat, k = 70)

clusters <- factor(membership(out[[2]]))

sce$cluster_70 <- clusters

unique(sce$cluster_70)

#Save SCEs after clustering
saveRDS(sce, "/mnt/central_nas/tnbc_volume/TNBC/data_analysis/06_sce_TNBC_TME_com.rds")

gc()

#k=90
out <- Rphenoannoy(mat, k = 90)

clusters <- factor(membership(out[[2]]))

sce$cluster_90 <- clusters

unique(sce$cluster_90)

#Save SCEs after clustering
saveRDS(sce, "/mnt/central_nas/tnbc_volume/TNBC/data_analysis/06_sce_TNBC_TME_com.rds")

gc()

#k=110
out <- Rphenoannoy(mat, k = 110)

clusters <- factor(membership(out[[2]]))

sce$cluster_110 <- clusters

unique(sce$cluster_110)

#Save SCEs after clustering
saveRDS(sce, "/mnt/central_nas/tnbc_volume/TNBC/data_analysis/06_sce_TNBC_TME_com.rds")

gc()

#k=130
out <- Rphenoannoy(mat, k = 130)

clusters <- factor(membership(out[[2]]))

sce$cluster_130 <- clusters

unique(sce$cluster_130)

#Save SCEs after clustering
saveRDS(sce, "/mnt/central_nas/tnbc_volume/TNBC/data_analysis/06_sce_TNBC_TME_com.rds")

gc()
```

## Cluster stability testing and comparison

```{r clustree, message = FALSE, warning=FALSE}
#Read sce after clustering and GMM
#sce <- readRDS("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/06_sce_TNBC_TME.rds")
#sce

#Pairwise rand index 
#Agreement between two clustering approaches - ARI > 0.5 = "good" similarity
pairwiseRand(sce$cluster_30,sce$cluster_50, mode = "index")
pairwiseRand(sce$cluster_50,sce$cluster_70, mode = "index")
pairwiseRand(sce$cluster_70,sce$cluster_90, mode = "index")
pairwiseRand(sce$cluster_90,sce$cluster_110, mode = "index")
pairwiseRand(sce$cluster_110,sce$cluster_130, mode = "index")

#Over different resolution parameters - Silhouette width [-1,1,max], Purity[0,1,max], Root-mean-squared difference [min]
clusters <- colnames(colData(sce))[str_detect(colnames(colData(sce)),"cluster_[:digit:]")]
df <- foreach(i = seq_along(clusters),.combine = rbind) %do% {
num_clus <- length(unique(colData(sce)[,clusters[i]]))
sil <- approxSilhouette(t(assay(sce, "exprs")), clusters=colData(sce)[,clusters[i]])
#purity <- neighborPurity(t(assay(sce, "exprs")), clusters=colData(sce)[,clusters[i]]) #takes very long to compute - do once and then save figure
rmsd <- clusterRMSD(t(assay(sce, "exprs")), clusters=colData(sce)[,clusters[i]])
data.frame(num_clus = num_clus,
           k = as.numeric(str_split(clusters[i],"_",simplify = TRUE)[,2]), 
           sil = mean(sil$width), 
#           purity = mean(purity$purity),
           rmsd = sum(rmsd))
}


p1 <- ggplot(df, aes(x=k, y=num_clus))+
  geom_point(aes(k, num_clus, col = num_clus))+
  geom_line(aes(k, num_clus,col=num_clus))+
  theme_classic()+
  scale_color_continuous(type = "viridis")

p2 <- ggplot(df, aes(x=k, y=rmsd))+
  geom_point(aes(k, rmsd, col = rmsd))+
  geom_line(aes(k, rmsd,col=rmsd))+
  theme_classic()+
  scale_color_continuous(type = "viridis")

p3 <- ggplot(df, aes(x=k, y=sil))+
  geom_point(aes(k, sil, col = sil))+
  geom_line(aes(k, sil,col=sil))+
  theme_classic()+
  scale_color_continuous(type = "viridis")
  
p1+p2+p3 

#Chosen method => k=130

#Cluster colors 
col_clus <- colorRampPalette(brewer.pal(9,"Dark2"))(length(unique(sce$cluster_130)))

#By clusters for one method (here k=130)
#1. Silhouette width [-1,1]
#Clusters with large positive silhouette widths are well-separated from other clusters
sil.approx <- approxSilhouette(t(assay(sce, "exprs")), clusters=sce$cluster_130)
sil.data <- as.data.frame(sil.approx)
sil.data$closest <- factor(ifelse(sil.data$width > 0, sce$cluster_130, sil.data$other))
sil.data$cluster <- sce$cluster_130

p4 <- ggplot(sil.data, aes(x=cluster, y=width, colour=closest)) +
    ggbeeswarm::geom_quasirandom(method="smiley")+
  theme_classic()+
  scale_color_manual(values = col_clus)
p4

# #By neighborpurity (takes very long to be computed - optional)
# pure.approx <- neighborPurity(t(assay(sce, "exprs")), clusters=sce$cluster_50)
# pure.data <- as.data.frame(pure.approx)
# pure.data$maximum <- factor(pure.data$maximum)
# pure.data$cluster <- sce$cluster_50
# 
# p5 <- ggplot(pure.data, aes(x=cluster, y=purity, colour=maximum)) +
#     ggbeeswarm::geom_quasirandom(method="smiley")+
#   theme_classic()+
#   scale_color_manual(values = col_clus)
# p5 
```

## Add missing metadata 

```{r}
# Missing metadata for 10 ZTMA174 patients
sce <- readRDS("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/06_sce_TNBC_TME_new.rds")

meta <- read.csv("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/03_data/Metadata_Combined_TNBC_LM.csv",row.names = 1)

## Add relevant metadata based on PID
for (entry in colnames(meta)[colnames(meta) %in% c("age","grade","pT", "pT_simple", "pN", "pN_simple", "status_OS","status_DFS", "OS_months","DFS_months")]) {
  colData(sce)[,entry] <- meta[,entry][match(colData(sce)[,"PID"],meta[,"PID"])]
}
```

```{r}
# Mark by pre-treatment information
sce$pretreatment_pT <- str_detect(sce$pT, "y")
sce$pretreatment_pN <- str_detect(sce$pN, "y")

test_df <- colData(sce) %>% as.data.frame %>% select(PID, pT, pT_simple, pN, pN_simple, pretreatment_pT, pretreatment_pN) %>% unique() %>% filter(pretreatment_pN == TRUE)

sce$pretreatment <- sce$pretreatment_pT

metadata(sce)$colors$pretreatment <- setNames(brewer.pal(4,"Set3")[3:4], sort(unique(sce$pretreatment)))
```

```{r}
# Proportion of pre-treatment per metacluster
pretreatment_mc <- colData(sce) %>% as.data.frame() %>% select(immune_metacluster,pretreatment) %>% group_by(immune_metacluster) %>% table() %>% as.data.frame
pretreatment_mc <- pretreatment_mc %>% group_by(immune_metacluster) %>% mutate(fra = Freq/sum(Freq)) 

ggplot(pretreatment_mc, aes(x = immune_metacluster, y = fra, fill = pretreatment))+
  geom_tile(position = "stack",color = "black")+#+alpha = 0.5)+
  theme_classic()+
  ylab("Fraction of cells per group")+
  scale_fill_manual(values = metadata(sce)$colors$pretreatment)
```

# 2. Heatmap visualization

## Heatmap with different scales 

```{r heatmap scales, message=FALSE,warning=FALSE}
#Heatmap body colors
col_exprs <- colorRamp2(c(0,1,2,3), c("#440154FF", "#31688EFF", "#35B779FF", "#FDE725FF"))
col_min_max <- colorRamp2(c(0,0.2,0.4,0.6,0.8), c("#440154FF", "#33628DFF", "#28AE80FF", "#DBE318FF", "#FDE725FF"))

### 1. Mean exprs counts
#Aggregrate across cells with the mean

cluster_mean_sce <- aggregateAcrossCells(sce, ids = sce$cluster_110,
                                 statistics = "mean",
                                 use.assay.type = "exprs",
                                 subset.row = rowData(sce)$cluster_channel)

cellstate_mean_sce <- aggregateAcrossCells(sce, ids = sce$cluster_110,
                                 statistics = "mean",
                                 use.assay.type = "exprs",
                                 subset.row = rowData(sce)$cell_state_channel)

other_mean_sce <- aggregateAcrossCells(sce, ids = sce$cluster_110,
                                 statistics = "mean",
                                 use.assay.type = "exprs",
                                 subset.row = rowData(sce)$other_epi_channel)

h_exprs <- Heatmap(t(assay(cluster_mean_sce, "exprs")),
        column_title = "1. mean exprs",
        col = col_exprs,
        name= "mean exprs",
        show_row_names = T, 
        show_column_names =  T, 
        row_names_side = "left",
        clustering_method_rows = "ward.D2",
        clustering_method_columns = "ward.D2")+
Heatmap(t(assay(cellstate_mean_sce, "exprs")),
        col = col_exprs,
        name= "mean exprs",
        show_row_names = T,
        show_column_names =  T,
        clustering_method_rows = "ward.D2",
        clustering_method_columns = "ward.D2")+
  Heatmap(t(assay(other_mean_sce, "exprs")),
        col = col_exprs,
        name= "mean exprs",
        show_row_names = T, 
        show_column_names =  T, 
        clustering_method_rows = "ward.D2",
        clustering_method_columns = "ward.D2")+
  rowAnnotation(ncells = anno_barplot(cluster_mean_sce$ncells))

h_exprs 

### 2. Mean min_max counts
#Aggregrate across cells with the mean
cluster_mean_sce <- aggregateAcrossCells(sce, ids = sce$cluster_130,
                                 statistics = "mean",
                                 use.assay.type = "min_max",
                                 subset.row = rowData(sce)$cluster_channel)

cellstate_mean_sce <- aggregateAcrossCells(sce, ids = sce$cluster_130,
                                 statistics = "mean",
                                 use.assay.type = "min_max",
                                 subset.row = rowData(sce)$cell_state_channel)

other_mean_sce <- aggregateAcrossCells(sce, ids = sce$cluster_130,
                                 statistics = "mean",
                                 use.assay.type = "min_max",
                                 subset.row = rowData(sce)$other_epi_channel)

h_min_max <- Heatmap(t(assay(cluster_mean_sce, "min_max")),
        column_title = "2. mean min_max",
        col = col_min_max,
        name= "min_max",
        show_row_names = T, 
        show_column_names =  T, 
        row_names_side = "left",
        clustering_method_rows = "ward.D2",
        clustering_method_columns = "ward.D2")+
Heatmap(t(assay(cellstate_mean_sce, "min_max")),
        col = col_min_max,
        name= "min_max",
        show_row_names = T,
        show_column_names =  T,
        clustering_method_rows = "ward.D2",
        clustering_method_columns = "ward.D2")+
Heatmap(t(assay(other_mean_sce, "min_max")),
        col = col_min_max,
        name= "min_max",
        show_row_names = T, 
        show_column_names =  T, 
        clustering_method_rows = "ward.D2",
        clustering_method_columns = "ward.D2")+
  rowAnnotation(ncells = anno_barplot(cluster_mean_sce$ncells))

h_exprs 
h_min_max 

h_exprs+
h_min_max
```

Observation: Mean "raw" or "min_max" exprs counts are most interpretable and logical - use raw exprs for downstream analysis

### Optional: Cluster merging of similar clusters with clear phenotype - NOT PERFORMED HERE

```{r cluster merg}
sce$cluster_50_merg <- sce$cluster_50
sce$cluster_50_merg <- ifelse(sce$cluster_50_merg %in% c("1","4"), "1_4", sce$cluster_50_merg)
sce$cluster_50_merg <- ifelse(sce$cluster_50_merg %in% c("2","9"), "2_9", sce$cluster_50_merg)
sce$cluster_50_merg <- as.factor(sce$cluster_50_merg)

cluster_50_merg_df <- data.frame(cluster_50_merg = unique(sce$cluster_50_merg), cluster_50_merg_names = 1:14) #define new and sorted cluster names
sce$cluster_50_renamed <- cluster_50_merg_df$cluster_50_merg_names[match(sce$cluster_50_merg,cluster_50_merg_df$cluster_50_merg)] 
sce$cluster_50_renamed <- as.factor(sce$cluster_50_renamed)
```

## Rich heatmap of clusters with additional information (other channels, spatial, grade, subtype)

```{r heatmap sce_1, message=FALSE,warning=FALSE}
### 1. Heatmap body ###

#Aggregrate across cells with the mean (median is also possible)
cluster_mean_sce <- aggregateAcrossCells(sce, ids = sce$cluster_130,
                                 statistics = "mean",
                                 use.assay.type = "exprs",
                                 subset.row = rowData(sce)$cluster_channel)

other_mean_sce <- aggregateAcrossCells(sce, ids = sce$cluster_130,
                                 statistics = "mean",
                                 use.assay.type = "exprs",
                                 subset.row = rowData(sce)$other_epi_channel)

# Heatmap body colors 
col_exprs <- colorRamp2(c(0,1,2,3), c("#440154FF", "#31688EFF", "#35B779FF", "#FDE725FF"))


### 2. Heatmap annotation ###

### 2.1 Spatial features

#Add number of neighbors to sce object (saved in ColPair)
n_neighbors <- colPair(sce) %>% as.data.frame %>% group_by(from) %>% dplyr::count() %>% arrange(desc(n))

sce$n_neighbors <- n_neighbors$n[match(seq_along(colnames(sce)),n_neighbors$from)]
sce$n_neighbors <- sce$n_neighbors %>% replace_na(0)

#Select spatial features and average over clusters 
spatial <- colData(sce) %>% as.data.frame %>% select(cluster_130, area, eccentricity, n_neighbors)
spatial <- spatial %>% select(-cluster_130) %>% aggregate(by = list(cluster_130 = spatial$cluster_130), FUN = mean) %>% column_to_rownames("cluster_130")

#Create Spatial Heatmap Annotation
ha_spatial <- HeatmapAnnotation(
    area = spatial$area,
    eccentricity = spatial$eccentricity,
    n_neighbors = spatial$n_neighbors, 
    col = list(area = metadata(sce)$colors$area,
               eccentricity = metadata(sce)$colors$eccentricity,
               n_neighbors = metadata(sce)$colors$n_neighbors),
    border = TRUE,
    which = "row")


### 2.2 Immune Metacluster 

anno <- colData(cluster_mean_sce) %>% as.data.frame %>% select(cluster_130, ncells)

anno <- anno %>% mutate(metacluster = ifelse(cluster_130 %in% c("4","10"),"CD8_T_cell",
                                     ifelse(cluster_130 %in% c("14","17","9"),"CD4_T_cell",
                                            #ifelse(cluster_130 %in% c("21"),"BnT_cell",
                                                   ifelse(cluster_130 %in% c("20","21"),"B_cell",
                                                          ifelse(cluster_130 %in% c("16","2","12"),"Macrophage",
                                                                 ifelse(cluster_130 %in% c("1"),"HLA_DR",
                                                                        ifelse(cluster_130 %in% c("19","18"),"Neutrophil",
                                                                               ifelse(cluster_130 %in% c("5","7","3"),"Endothelial",
                                                                                      ifelse(cluster_130 %in% c("13","8","6","11"),"Stromal",
                                                                                             ifelse(cluster_130 %in% c("15"),"Immune_low",NA)
                                                                                             ))))))))) %>% mutate(metacluster = as.factor(metacluster))

sce$immune_metacluster <- anno$metacluster[match(sce$cluster_130, anno$cluster_130)]


### 2.3 Astir immune cell types 

#Export arsinh normalized expression matrix as csv
# Add sampleID as covariate in design matrix

expr_astir <- assay(sce,"exprs") %>% t() %>% as.data.frame() %>% mutate(sample = sce$sample_id)

#write.csv(expr_astir, file = "/mnt/central_nas/tnbc_volume/TNBC/data_analysis/04_astir/expr_TNBC_TME_astir.csv",row.names = TRUE)

### RUN ASTIR IN CONDA_ENVIRONMENT ###

# Import Astir results 

astir_type <- read_csv("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/04_astir/cell_types_TNBC_TME.csv",col_names = c("ID","cell_type"),skip = 1)

astir_sum <- astir_type %>% group_by(cell_type) %>% count()

sce$astir_TME <- astir_type$cell_type

## Create ASTIR annotation
# Astir cell types per cluster
astir_anno <- colData(sce) %>% as.data.frame() %>% select(cluster_130,astir_TME) %>% group_by(cluster_130) %>% table() %>% as.data.frame
astir_anno <- astir_anno %>% group_by(cluster_130) %>% mutate(fra = Freq/sum(Freq)) 
astir_anno <- astir_anno %>% select(-Freq) %>% pivot_wider(id_cols = cluster_130,names_from = astir_TME,values_from = fra) %>% column_to_rownames("cluster_130")

#Create colors for astir cell types 
astir_col <- setNames(c("#7FC97F","#E1B8A8","#FDD58C","#E8EE9B","#DA2950","#666666","white"),sort(unique(sce$astir_TME)))

metadata(sce)$colors$astir_col <- astir_col

ha_astir <- HeatmapAnnotation(astir_TME = anno_barplot(astir_anno, width = unit(20, "mm"), gp = gpar(fill = metadata(sce)$colors$astir_col)),
                            border = TRUE, 
                            annotation_name_rot = 90,
                            gap = unit(3,"mm"),
                            which = "row")

###2.3 Metadata features

#Number of cells per cluster 
n_cells <- colData(cluster_mean_sce) %>% as.data.frame %>% select(cluster_130, ncells)

# Proportion of grade per cluster
grade <- colData(sce) %>% as.data.frame() %>% select(cluster_130,grade) %>% group_by(cluster_130) %>% table() %>% as.data.frame
grade <- grade %>% group_by(cluster_130) %>% mutate(fra = Freq/sum(Freq)) 
grade <- grade %>% select(-Freq) %>% pivot_wider(id_cols = cluster_130,names_from = grade,values_from = fra) %>% column_to_rownames("cluster_130")

# Proportion of pT stage per cluster
pT <- colData(sce) %>% as.data.frame() %>% select(cluster_130,pT_simple) %>% group_by(cluster_130) %>% table() %>% as.data.frame
pT <- pT %>% group_by(cluster_130) %>% mutate(fra = Freq/sum(Freq)) 
pT <- pT %>% select(-Freq) %>% pivot_wider(id_cols = cluster_130,names_from = pT_simple,values_from = fra) %>% column_to_rownames("cluster_130")

# Proportion of pN stage per cluster
pN <- colData(sce) %>% as.data.frame() %>% select(cluster_130,pN_simple) %>% group_by(cluster_130) %>% table() %>% as.data.frame
pN <- pN %>% group_by(cluster_130) %>% mutate(fra = Freq/sum(Freq)) 
pN <- pN %>% select(-Freq) %>% pivot_wider(id_cols = cluster_130,names_from = pN_simple,values_from = fra) %>% column_to_rownames("cluster_130")

# Number of contributing patients
cluster_PID <- colData(sce) %>% as.data.frame() %>% select(cluster_130,PID) %>% group_by(cluster_130) %>% table() %>% as.data.frame
n_PID <- cluster_PID %>% filter(Freq>0) %>% group_by(cluster_130) %>% count(name = "n_PID") %>% column_to_rownames("cluster_130")


# Define color scheme for immune clusters/metaclusters
row_order_clusters <- row_order(Heatmap(t(assay(cluster_mean_sce, "exprs")),
        column_title = "cluster_channels",
        cluster_row_slices = TRUE,
        clustering_method_rows = "ward.D2",
        clustering_method_columns = "ward.D2"))


cluster_col <- setNames(colorRampPalette(brewer.pal(9, "Accent"))(length(unique(anno$cluster_130))), row_order_clusters)                                        
metacluster_col <- setNames(colorRampPalette(brewer.pal(9, "Accent"))(length(unique(anno$metacluster))), sort(unique(anno$metacluster)))                                          
metadata(sce)$colors$cluster_130 <- cluster_col
metadata(sce)$colors$immune_metacluster <- metacluster_col

# Metadata annotation
ha_meta <- HeatmapAnnotation(cluster_130 = anno_simple(unfactor(anno$cluster_130), width = unit(8, "mm"), pch = unfactor(anno$cluster_130),border=TRUE,gp = gpar(fill = "white")),
                            metacluster = anno_simple(unfactor(anno$metacluster),border=TRUE,col=metadata(sce)$colors$immune_metacluster),
                            n_cells = anno_barplot(n_cells$ncells, width = unit(10, "mm"),gp = gpar(fill = "#440154FF")),
                            n_PID = anno_barplot(n_PID, width = unit(10, "mm"),gp = gpar(fill = "#440154FF")),
                            grade = anno_barplot(grade, width = unit(12, "mm"),gp = gpar(fill = metadata(sce)$colors$grade)),
                            pT = anno_barplot(pT, width = unit(12, "mm"),gp = gpar(fill = metadata(sce)$colors$pT_simple)),
                            pN = anno_barplot(pN, width = unit(12, "mm"),gp = gpar(fill = metadata(sce)$colors$pN_simple)),
                            border = TRUE, 
                            annotation_name_rot = 90,
                            gap = unit(1,"mm"),
                            which = "row")

### 3. Plot heatmap ###

h <- Heatmap(t(assay(cluster_mean_sce, "exprs")),
        column_title = "cluster_channels",
        col = col_exprs,
        name = "mean exprs",
        row_title = NULL, 
        show_column_names =  TRUE, 
        clustering_method_rows = "ward.D2",
        clustering_method_columns = "ward.D2",
        row_split = anno$metacluster,
        cluster_row_slices = FALSE)+
  # Heatmap(t(assay(other_mean_sce, "exprs")),
  #       column_title = "other_channels",
  #       col = col_exprs,
  #       name= "mean exprs",
  #       show_row_names = FALSE, 
  #       show_column_names =  TRUE, 
  #       clustering_method_rows = "ward.D2",
  #       clustering_method_columns = "ward.D2")+
  ha_astir+
  ha_meta+
  ha_spatial

# Add customized legend for anno_barplot()
lgd_astir <- Legend(title = "Astir_TME", at = names(metadata(sce)$colors$astir_col), 
              legend_gp = gpar(fill = metadata(sce)$colors$astir_col))

lgd_meta <- Legend(title = "Immune_Metacluster", at = names(metadata(sce)$colors$immune_metacluster), 
              legend_gp = gpar(fill = metadata(sce)$colors$immune_metacluster))          
# Plot
draw(h,annotation_legend_list = list(lgd_astir, lgd_meta))
```

```{r save_objects, message = FALSE, warning=FALSE}
#Save sce
#saveRDS(sce, "/mnt/central_nas/tnbc_volume/TNBC/data_analysis/06_sce_TNBC_TME_com.rds")
```


# 3. Single-cell visualization

## Dimensionality reduction 

```{r dimred, message=FALSE,warning=FALSE}
## Subsample 10% of cells from each core 
#to save time and not get super crowded dimensionality reduction plots
set.seed(22)
sub <- colData(sce) %>% as.data.frame() %>% group_by("sample_id") %>% slice_sample(prop = 0.1) %>% arrange("sample_id") %>% mutate(sub_id = paste(sample_id,ObjectNumber,sep="_")) %>% pull(sub_id)
sce_sub = sce[,sub]

#Run UMAP/TSNE on cluster channels
sce_sub = runUMAP(sce_sub[rowData(sce_sub)$cluster_channel,], n_neighbors = 50, pca = 50, exprs_values = "exprs",external_neighbors=TRUE, BPPARAM = MulticoreParam((detectCores()-2)))
sce_sub = runTSNE(sce_sub[rowData(sce_sub)$cluster_channel,], exprs_values = "exprs", external_neighbors=TRUE, BPPARAM = MulticoreParam((detectCores()-2)))


## 1. Visualize UMAP and TSNE
#panCK and GMM_CK
UMAP_clus <- plotReducedDim(sce_sub, dimred = "UMAP", colour_by = "cluster_50_renamed")+
  scale_color_manual(values = col_clusters$cluster_50_renamed)

TSNE_clus <- plotReducedDim(sce_sub, dimred = "TSNE", colour_by = "cluster_50_renamed")+
  scale_color_manual(values = col_clusters$cluster_50_renamed)

(UMAP_clus+TSNE_clus)

#Visualize all marker expression
all_plots_UMAP <- lapply(c(rownames(sce_sub)[rowData(sce_sub)$cluster_channel]),
                    function(x){
                      p <- plotReducedDim(sce_sub, dimred = "UMAP", 
                                          colour_by = x, 
                                          by_exprs_values = "exprs",
                                          point_size = 0.5)
                      return(p)                    
                    })

library(cowplot)
exp_plots_UMAP <- plot_grid(plotlist = all_plots_UMAP)

all_plots_TSNE <- lapply(c(rownames(sce_sub)[rowData(sce_sub)$cluster_channel]),
                    function(x){
                      p <- plotReducedDim(sce_sub, dimred = "TSNE", 
                                          colour_by = x, 
                                          by_exprs_values = "exprs",
                                          point_size = 0.5)
                      return(p)                    
                    })

library(cowplot)
exp_plots_TSNE <- plot_grid(plotlist = all_plots_TSNE)

#Save subsampled sce
#saveRDS(sce_sub,"/mnt/central_nas/tnbc_volume/TNBC/data_analysis/sce_sub_TNBC_249_TME_CC_ClusterChannelClus_k.rds")
```

## Heatmap for subset of cells

```{r heatmap sub}
#Heatmap visualization
dittoHeatmap(sce_sub, 
             genes = rownames(sce_sub)[rowData(sce_sub)$cluster_channel],
             assay = "exprs", 
             scale = "none", 
             order.by = c("immune_metacluster", "cluster_130"), 
             heatmap.colors = viridis(100), 
             annot.by = c("immune_metacluster", "cluster_130"),
             annotation_colors = list(immune_metacluster = metadata(sce_sub)$colors$immune_metacluster, 
                                      cluster_130 = metadata(sce_sub)$colors$cluster_130)
)
             
```

## Violin plot visualization

```{r celltype violin, message=FALSE, fig.width=6, fig.height=10}
#Violin Plot - plotExpression
plotExpression(sce_sub, 
               features = rownames(sce_sub)[rowData(sce_sub)$cluster_channel],
               x = "immune_metacluster", exprs_values = "exprs", 
               colour_by = "immune_metacluster") +
    theme(axis.text.x =  element_text(angle = 90))+
    scale_color_manual(values = metadata(sce_sub)$colors$immune_metacluster)
```

## Barplot visualization 

### Fractions - Visualization and correlation of cluster fractions

```{r cluster-pro, message = FALSE, warning = FALSE}
## 1. Fraction of immune_immune_metacluster/cluster per sample/PID
dittoBarPlot(sce, var = "immune_metacluster", group.by = "sample_id")+
  scale_fill_manual(values = metadata(sce)$colors$immune_metacluster)+
  theme(axis.text.x = element_blank())

d_PID <- dittoBarPlot(sce, var = "immune_metacluster", group.by = "PID")+
  scale_fill_manual(values = metadata(sce)$colors$immune_metacluster)+
  theme(axis.text.x = element_blank())

dittoBarPlot(sce, var = "cluster_130", group.by = "sample_id")+
  scale_fill_manual(values = metadata(sce)$colors$cluster_130)+
  theme(axis.text.x = element_blank())

dittoBarPlot(sce, var = "cluster_130", group.by = "PID")+
  scale_fill_manual(values = metadata(sce)$colors$cluster_130)+
  theme(axis.text.x = element_blank())

## 2. Fraction of sample/PID per immune_metacluster/cluster
dittoBarPlot(sce, var = "PID", group.by = "immune_metacluster")+
  scale_fill_manual(values = metadata(sce)$colors$PID)+
  guides(fill = "none")

dittoBarPlot(sce, var = "PID", group.by = "cluster_130")+
  scale_fill_manual(values = metadata(sce)$colors$PID)+
  guides(fill = "none")

dittoBarPlot(sce, var = "sample_id", group.by = "immune_metacluster")+
  guides(fill = "none")

dittoBarPlot(sce, var = "sample_id", group.by = "cluster_130")+
  guides(fill = "none")


## 3. Correlation of immune_metacluster/cluster fractions for each PID (2-4 samples per PID)
SID_cluster <- colData(sce) %>% as.data.frame() %>% select(cluster_130,sample_id) %>% group_by(sample_id) %>% table() %>% as.data.frame
SID_cluster <- SID_cluster %>% group_by(sample_id) %>% mutate(fra = Freq/sum(Freq))

SID_immune_metacluster <- colData(sce) %>% as.data.frame() %>% select(immune_metacluster,sample_id) %>% group_by(sample_id) %>% table() %>% as.data.frame
SID_immune_metacluster <- SID_immune_metacluster %>% group_by(sample_id) %>% mutate(fra = Freq/sum(Freq))

# Cluster fraction correlation for each PID (2 samples per PID)
## by cluster
SID_cluster_cor <- SID_cluster
SID_cluster_cor$PID <- sce$PID[match(SID_cluster$sample_id,sce$sample_id)]

SID_cluster_cor <- foreach(i = unique(colData(sce)$PID),.combine = "rbind")%do%{
SID <- SID_cluster_cor %>% filter(PID %in% i) %>% select(-Freq,-PID) %>% pivot_wider(id_cols = cluster_130,names_from = sample_id,values_from = fra) %>% column_to_rownames("cluster_130")
cor <- cor(SID,method = "pearson")
cor[2]
}

SID_cluster_cor <- as.data.frame(SID_cluster_cor) 
SID_cluster_cor$PID <- unique(colData(sce)$PID) 
SID_cluster_cor$category <- "PID"

SID_cluster_cor_plot <- ggplot(SID_cluster_cor, aes(x = category, y = V1))+
  geom_boxplot(outlier.colour = NA)+
  geom_jitter(aes(color = PID), size = 3, position=position_jitter(w=0.1,h=0.1))+
  guides(color = "none")+
  scale_color_manual(values = metadata(sce)$colors$PID)+
  theme_classic(base_size = 12)+
  theme(axis.title.x = element_blank())+
  ylab("Correlation of cluster fractions per PID (n=2-4)")+
  ggtitle(paste("Spearman rho = ",round(median(SID_cluster_cor$V1, na.rm = TRUE), digits = 3)))

SID_cluster_cor_plot

## by immune_metacluster
SID_immune_metacluster_cor <- SID_immune_metacluster
SID_immune_metacluster_cor$PID <- sce$PID[match(SID_immune_metacluster$sample_id,sce$sample_id)]

SID_immune_metacluster_cor <- foreach(i = unique(colData(sce)$PID),.combine = "rbind")%do%{
SID <- SID_immune_metacluster_cor %>% filter(PID %in% i) %>% select(-Freq,-PID) %>% pivot_wider(id_cols = immune_metacluster,names_from = sample_id,values_from = fra) %>% column_to_rownames("immune_metacluster")
cor <- cor(SID,method = "spearman")
cor[2]
}

SID_immune_metacluster_cor <- as.data.frame(SID_immune_metacluster_cor) 
SID_immune_metacluster_cor$PID <- unique(colData(sce)$PID) 
SID_immune_metacluster_cor$category <- "PID"

SID_immune_metacluster_cor_plot <- ggplot(SID_immune_metacluster_cor, aes(x = category, y = V1))+
  geom_boxplot(outlier.colour = NA)+
  geom_jitter(aes(color = PID), size = 3, position=position_jitter(w=0.1,h=0.1))+
  guides(color = "none")+
  scale_color_manual(values = metadata(sce)$colors$PID)+
  theme_classic(base_size = 12)+
  theme(axis.title.x = element_blank())+
  ylab("Correlation of immune_metacluster fractions per PID (n=2-4)")+
  ggtitle(paste("Spearman rho = ",round(median(SID_immune_metacluster_cor$V1, na.rm = TRUE), digits = 2)))


SID_immune_metacluster_cor_plot

d_PID + SID_immune_metacluster_cor_plot
```

### Densities - Visualization and correlation of cluster densities

```{r density cor}
# 1. By metaclusters

## Density barplot

### For each sample
SID_immune_metacluster <- colData(sce) %>% as.data.frame() %>% select(immune_metacluster,sample_id) %>% group_by(sample_id) %>% table() %>% as.data.frame

SID_immune_metacluster$tissue_area <- sce$tissue_area[match(SID_immune_metacluster$sample_id,sce$sample_id)]

SID_immune_metacluster <- SID_immune_metacluster %>% mutate(density = Freq/(tissue_area/1000000))

ggplot(SID_immune_metacluster) +
  geom_tile(color = "black",aes(sample_id, density, fill = immune_metacluster), position = "stack")+
  scale_fill_manual(values=metadata(sce)$colors$immune_metacluster) + 
  theme_classic(base_size = 15)+
  ylab("Density")+
  xlab("PID")+
  theme(axis.text.x = element_blank())


### For each PID
PID_immune_metacluster <- colData(sce) %>% as.data.frame() %>% select(immune_metacluster,PID) %>% group_by(PID) %>% table() %>% as.data.frame

PID_immune_metacluster$tissue_area <- sce$tissue_area[match(PID_immune_metacluster$PID,sce$PID)]

PID_immune_metacluster <- PID_immune_metacluster %>% mutate(density = Freq/(tissue_area/1000000))

d_PID_dens <- ggplot(PID_immune_metacluster) +
  geom_tile(color = "black",aes(PID, density, fill = immune_metacluster), position = "stack")+
  scale_fill_manual(values=metadata(sce)$colors$immune_metacluster) + 
  theme_classic(base_size = 15)+
  ylab("Density")+
  xlab("PID")+
  theme(axis.text.x = element_blank())


## Correlation of immune_metacluster/cluster densities for each PID (2-4 samples per PID) - SAME AS FOR FRACTION CORRELATION?

### By metacluster density
SID_metacluster_cor <- SID_immune_metacluster

SID_metacluster_cor$PID <- sce$PID[match(SID_metacluster_cor$sample_id,sce$sample_id)]

SID_metacluster_cor <- foreach(i = unique(colData(sce)$PID),.combine = "rbind")%do%{
SID <- SID_metacluster_cor %>% filter(PID %in% i) %>% select(-Freq,-PID) %>% pivot_wider(id_cols = immune_metacluster,names_from = sample_id,values_from = density) %>% column_to_rownames("immune_metacluster")
cor <- cor(SID,method = "spearman")
cor[2]
}

SID_metacluster_cor <- as.data.frame(SID_metacluster_cor) 
SID_metacluster_cor$PID <- unique(colData(sce)$PID) 
SID_metacluster_cor$category <- "PID"

SID_metacluster_dens_cor_plot <- ggplot(SID_metacluster_cor, aes(x = category, y = V1))+
  geom_boxplot(outlier.colour = NA)+
  geom_jitter(aes(color = PID), size = 3, position=position_jitter(w=0.1,h=0.1))+
  guides(color = "none")+
  scale_color_manual(values = metadata(sce)$colors$PID)+
  theme_classic(base_size = 12)+
  theme(axis.title.x = element_blank())+
  ylab("Correlation of cluster fractions per PID (n=2-4)")+
  ggtitle(paste("Spearman rho = ",round(median(SID_metacluster_cor$V1, na.rm = TRUE), digits = 2)))

SID_metacluster_dens_cor_plot


### By cluster density
SID_immune_cluster <- colData(sce) %>% as.data.frame() %>% select(cluster_130,sample_id) %>% group_by(sample_id) %>% table() %>% as.data.frame

SID_immune_cluster$tissue_area <- sce$tissue_area[match(SID_immune_cluster$sample_id,sce$sample_id)]

SID_immune_cluster <- SID_immune_cluster %>% mutate(density = Freq/(tissue_area/1000000))

SID_cluster_dens_cor <- SID_immune_cluster

SID_cluster_dens_cor$PID <- sce$PID[match(SID_cluster_dens_cor$sample_id,sce$sample_id)]

SID_cluster_dens_cor <- foreach(i = unique(colData(sce)$PID),.combine = "rbind")%do%{
SID <- SID_cluster_dens_cor %>% filter(PID %in% i) %>% select(-Freq,-PID,-tissue_area) %>% pivot_wider(id_cols = cluster_130,names_from = sample_id,values_from = density) %>% column_to_rownames("cluster_130")
cor <- cor(SID,method = "pearson")
cor[2]
}

SID_cluster_dens_cor <- as.data.frame(SID_cluster_dens_cor) 
SID_cluster_dens_cor$PID <- unique(colData(sce)$PID) 
SID_cluster_dens_cor$category <- "PID"

SID_cluster_dens_cor_plot <- ggplot(SID_cluster_dens_cor, aes(x = category, y = V1))+
  geom_boxplot(outlier.colour = NA)+
  geom_jitter(aes(color = PID), size = 3, position=position_jitter(w=0.1,h=0.1))+
  guides(color = "none")+
  scale_color_manual(values = metadata(sce)$colors$PID)+
  theme_classic(base_size = 12)+
  theme(axis.title.x = element_blank())+
  ylab("Correlation of cluster fractions per PID (n=2-4)")+
  ggtitle(paste("Spearman rho = ",round(median(SID_cluster_cor$V1, na.rm = TRUE), digits = 3)))

(d_PID + SID_immune_metacluster_cor_plot)/
(d_PID_dens + SID_metacluster_dens_cor_plot)
```


### Complex Heatmap based fraction stacked barplot

```{r}
## PID immune_metacluster fraction dataframe
PID_immune_metacluster <- colData(sce) %>% as.data.frame() %>% select(immune_metacluster,PID) %>% group_by(PID) %>% table() %>% as.data.frame
PID_immune_metacluster <- PID_immune_metacluster %>% group_by(PID) %>% mutate(fra = Freq/sum(Freq))
cur_df <- PID_immune_metacluster %>% pivot_wider(id_cols = "PID", names_from = "immune_metacluster",values_from = "fra") %>% column_to_rownames("PID")

# Perform hierarchical clustering on fractions
dd <- dist(cur_df, method = "euclidean")
hc <- hclust(dd, method = "ward.D2")
plot(hc)
row_sorted <- hc$order

# Define annotation dataframe
cur_df <- cur_df[row_sorted,]

anno <- colData(sce) %>% as.data.frame %>% select(PID, age_bin, grade, pT_simple, pN_simple, pM_simple, pretreatment) %>% unique()
anno <- anno[match(rownames(cur_df), anno$PID),]

# Plot annotation dataframe with ComplexHeatmap
fraction_anno <- HeatmapAnnotation(MC_fractions = anno_barplot(cur_df, gp=gpar(fill=metadata(sce)$colors$immune_metacluster[colnames(cur_df)]),
                                   bar_width = 1,
                                   height = unit(25,"cm"),
                                   width = unit(11,"cm"),
                                   show_row_names = FALSE), 
                         # PID = anno_simple(anno$PID, col=metadata(sce)$colors$PID, border=TRUE),
                          age = anno_simple(anno$age_bin, col = metadata(sce)$colors$age_bin, border = TRUE),
                          grade = anno_simple(anno$grade, col = metadata(sce)$colors$grade, border = TRUE),
                          pT = anno_simple(anno$pT_simple, col=metadata(sce)$colors$pT_simple, border=TRUE),
                          pN = anno_simple(anno$pN_simple, col=metadata(sce)$colors$pN_simple, border=TRUE),
                          pM = anno_simple(anno$pM_simple, col=metadata(sce)$colors$pM_simple, border = TRUE),
                          pretreatment = anno_simple(as.character(anno$pretreatment), col=metadata(sce)$colors$pretreatment, border = TRUE),
                          border = TRUE, 
                          #annotation_name_rot = 90,
                          gap = unit(1,"mm"),
                          which = "row")

fraction_anno + NULL
```





## CATALYST-based visualization

We will first convert the sce into a CATALYST-compatible format.

```{r CATALYST}
library(CATALYST)

# save spe in CATALYST-compatible object with renamed colData entries and 
# new metadata information
sce_cat <- sce 
sce_cat$sample_id <- factor(sce$PID)
sce_cat$cluster_id <- factor(sce$immune_metacluster)

#add celltype information to metadata
metadata(sce_cat)$cluster_codes <- data.frame(immune_metacluster = factor(sce_cat$immune_metacluster))
```

```{r CATALYST pbmds and clrDR, message=FALSE}
# MDS pseudobulk by metacluster
pbMDS(sce_cat, by = "cluster_id", 
      features = rownames(sce_cat)[rowData(sce_cat)$cluster_channel], 
      label_by = "cluster_id", k = "immune_metacluster") +
  scale_color_manual(values = metadata(sce_cat)$colors$immune_metacluster)+
  theme_classic()

# MDS pseudobulk by metacluster
pbMDS(sce_cat, by = "both", 
      features = rownames(sce_cat)[rowData(sce_cat)$cluster_channel], 
      k = "immune_metacluster",
      size_by = TRUE) +
  scale_color_manual(values = metadata(sce_cat)$colors$immune_metacluster)+
  theme_classic()

# CLR on metacluster proportions across samples
clrDR(sce_cat, dr = "PCA", 
      by = "cluster_id", k = "immune_metacluster", 
      label_by = "cluster_id", arrow_col = "PID", 
      point_pal = metadata(sce_cat)$colors$immune_metacluster, 
      arrows = FALSE) +
  scale_color_manual(values = metadata(sce_cat)$colors$PID)+
  guides(color = "none")+
  theme_classic()

# CLR on sample proportions across metaclusters
clrDR(sce_cat, dr = "PCA", 
      by = "sample_id", point_col = "sample_id",
      k = "immune_metacluster", point_pal = metadata(sce_cat)$colors$PID)+
  scale_color_manual(values = metadata(sce_cat)$colors$immune_metacluster)+
  guides(fill = "none")+
  theme_classic()
```

## Check correlation of markers

```{r marker correlations}
# Cluster channel marker correlation (all markers - does not show distribution which is usually recommended)
library(corrplot)

cor_sce <- as.data.frame(t(assay(sce,"exprs")[rowData(sce)$cluster_channel,]))

corrplot(cor(cor_sce,method = "spearman"), method="color", col=brewer.pal(n=8, name="RdBu"),  
         type="upper", order="hclust",hclust.method = "ward.D2", 
         addCoef.col = "black", addCoefasPercent = F,# Add coefficient of correlation
         tl.col="black", tl.srt=40, #Text label color and rotation
         # Combine with significance
         #p.mat = p.mat, sig.level = 0.01, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag=F
)

# For specific markers of interest (also shows the value distribution which is recommended)
#pairs.panels(cor_sce[,c("CD3","CD8a")],
#             method = "spearman",
#             hist.col = "royalblue1",
#             digits = 2)
```
# 4. Image visualization

```{r load images, message = FALSE}
#Load images and masks
images <- readRDS("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/images_segtest.rds")
masks <- readRDS("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/masks_segtest_4.rds")
```

## Visualize clusters on images 

```{r viz clus, message=FALSE, warning=FALSE}
# 1. Part 1 of cluster viz 
clus_select <- anno %>% filter(!metacluster %in% c("Endothelial","Stromal","Neutrophil")) %>% pull(cluster_130) %>% unfactor()

for (i in clus_select) {
  
##Visualize clusters on images - select 10 images for each category with highest number of cells
cur_sample <- colData(sce) %>% as.data.frame %>% group_by(cluster_130) %>% filter(cluster_130 == i) %>% count(sample_id) %>% arrange(desc(n)) %>% slice(1:10) %>% pull(sample_id)

#Prepare images
cur_images <- images[cur_sample]
cur_images <- cytomapper::normalize(cur_images, separateImages = TRUE)
cur_images <- cytomapper::normalize(cur_images, inputRange = c(0, 0.2))

#Subset sce
cur_sce <- sce[,sce$cluster_130 == i]

#Plot 
plotPixels(cur_images,
           mask = masks[cur_sample],
           object = cur_sce,
           img_id = "sample_id",
           cell_id = "ObjectNumber",
           missing_colour = "white",
           outline_by = "cluster_130",
           colour_by = c("CD4", "HLADR", "CD68","CD8a", "CD20","CD3"),
           #colour = list("SMA" = c("black", "green"),
           #              "CD3" = c("black", "yellow"),
           #              "CD31_vWF" = c("black", "magenta"),
           #              "CD20" = c("black", "cyan"),
           #              "CD68" = c("black", "red"),
           #              "cluster_130" = metadata(test_sce)$colors$cluster_130
           #              ),
           thick = FALSE,
           image_title = NULL,
           legend = list(colour_by.title.cex = 0.7,
                         colour_by.labels.cex = 0.7),
           save_plot = list(filename = paste0("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/TME_clusters_viz/TME_viz_cluster_",i,".png"), scale = 2)
           )
 
  }


# 2. Part 2 of cluster viz 

clus_select <- anno %>% filter(metacluster %in% c("Endothelial","Stromal","Neutrophil")) %>% pull(cluster_130) %>% unfactor()

for (i in clus_select) {
  
##Visualize clusters on images - select 10 images for each category with highest number of cells
cur_sample <- colData(sce) %>% as.data.frame %>% group_by(cluster_130) %>% filter(cluster_130 == i) %>% count(sample_id) %>% arrange(desc(n)) %>% slice(1:10) %>% pull(sample_id)

#Prepare images
cur_images <- images[cur_sample]
cur_images <- cytomapper::normalize(cur_images, separateImages = TRUE)
cur_images <- cytomapper::normalize(cur_images, inputRange = c(0, 0.2))

#Subset sce
cur_sce <- sce[,sce$cluster_130 == i]

#Plot 
plotPixels(cur_images,
           mask = masks[cur_sample],
           object = cur_sce,
           img_id = "sample_id",
           cell_id = "ObjectNumber",
           missing_colour = "white",
           outline_by = "cluster_130",
           colour_by = c("CD31_vWF", "SMA", "CD68","CD15", "CD20","CD3"),
           #colour = list("SMA" = c("black", "green"),
           #              "CD3" = c("black", "yellow"),
           #              "CD31_vWF" = c("black", "magenta"),
           #              "CD20" = c("black", "cyan"),
           #              "CD68" = c("black", "red"),
           #              "cluster_130" = metadata(test_sce)$colors$cluster_130
           #              ),
           thick = FALSE,
           image_title = NULL,
           legend = list(colour_by.title.cex = 0.7,
                         colour_by.labels.cex = 0.7),
           save_plot = list(filename = paste0("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/TME_clusters_viz/TME_viz_cluster_",i,".png"), scale = 2)
           )
 
  }
```


## Visualize metaclusters on images

```{r viz metaclus, message=FALSE, warning=FALSE}

# 1. Part 1 of immune_metacluster viz 
clus_select <- anno %>% filter(!metacluster %in% c("Endothelial","Stromal","Neutrophil")) %>% pull(metacluster) %>% unfactor()

for (i in clus_select) {
  
##Visualize clusters on images - select 10 images for each category with highest number of cells
cur_sample <- colData(sce) %>% as.data.frame %>% group_by(immune_metacluster) %>% filter(immune_metacluster == i) %>% count(sample_id) %>% arrange(desc(n)) %>% slice(1:10) %>% pull(sample_id)

#Prepare images
cur_images <- images[cur_sample]
cur_images <- cytomapper::normalize(cur_images, separateImages = TRUE)
cur_images <- cytomapper::normalize(cur_images, inputRange = c(0, 0.2))

#Subset sce
cur_sce <- sce[,sce$immune_metacluster == i]

#Plot 
plotPixels(cur_images,
           mask = masks[cur_sample],
           object = cur_sce,
           img_id = "sample_id",
           cell_id = "ObjectNumber",
           missing_colour = "white",
           outline_by = "cluster_130",
           colour_by = c("CD4", "HLADR", "CD68","CD8a", "CD20","CD3"),
           #colour = list("SMA" = c("black", "green"),
           #              "CD3" = c("black", "yellow"),
           #              "CD31_vWF" = c("black", "magenta"),
           #              "CD20" = c("black", "cyan"),
           #              "CD68" = c("black", "red"),
           #              "cluster_130" = metadata(test_sce)$colors$cluster_130
           #              ),
           thick = FALSE,
           image_title = NULL,
           legend = list(colour_by.title.cex = 0.7,
                         colour_by.labels.cex = 0.7),
           save_plot = list(filename = paste0("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/TME_clusters_viz/TME_viz_immune_metacluster_",i,".png"), scale = 2)
           )
 
  }


# 2. Part 2 of immune_metacluster viz 

clus_select <- anno %>% filter(metacluster %in% c("Endothelial","Stromal","Neutrophil")) %>% pull(metacluster) %>% unfactor()

for (i in clus_select) {
  
##Visualize clusters on images - select 10 images for each category with highest number of cells
cur_sample <- colData(sce) %>% as.data.frame %>% group_by(immune_metacluster) %>% filter(immune_metacluster == i) %>% count(sample_id) %>% arrange(desc(n)) %>% slice(1:10) %>% pull(sample_id)

#Prepare images
cur_images <- images[cur_sample]
cur_images <- cytomapper::normalize(cur_images, separateImages = TRUE)
cur_images <- cytomapper::normalize(cur_images, inputRange = c(0, 0.2))

#Subset sce
cur_sce <- sce[,sce$immune_metacluster == i]

#Plot 
plotPixels(cur_images,
           mask = masks[cur_sample],
           object = cur_sce,
           img_id = "sample_id",
           cell_id = "ObjectNumber",
           missing_colour = "white",
           outline_by = "cluster_130",
           colour_by = c("CD31_vWF", "SMA", "CD68","CD15", "CD20","CD3"),
           #colour = list("SMA" = c("black", "green"),
           #              "CD3" = c("black", "yellow"),
           #              "CD31_vWF" = c("black", "magenta"),
           #              "CD20" = c("black", "cyan"),
           #              "CD68" = c("black", "red"),
           #              "cluster_130" = metadata(test_sce)$colors$cluster_130
           #              ),
           thick = FALSE,
           image_title = NULL,
           legend = list(colour_by.title.cex = 0.7,
                         colour_by.labels.cex = 0.7),
           save_plot = list(filename = paste0("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/TME_clusters_viz/TME_viz_immune_metacluster_",i,".png"), scale = 2)
           )
 
  }
```


# 5. Survival analysis

## Cox-proportonal hazards model for metaclusters

### For OS on metaclusters

```{r cox}
library(survival)
library(survminer)

PID_metacluster <- colData(sce) %>% as.data.frame() %>% select(immune_metacluster,PID) %>% group_by(PID) %>% table() %>% as.data.frame
PID_metacluster <- PID_metacluster %>% group_by(PID) %>% mutate(fra = Freq/sum(Freq))

### Multivariate cox proportional hazard model on multiple covariates (here proportion of phenotypes)
PID_metagroups <- PID_metacluster %>% pivot_wider(id_cols = "PID",names_from = "immune_metacluster",values_from = "fra") %>% column_to_rownames("PID")
#colnames(PID_metagroups) <- paste0("MC",colnames(PID_metagroups))

## 1. Based on continuous cluster proportions
Cox_df <- PID_metagroups

#add covariates for correction
Cox_df$status_OS <- sce$status_OS[match(rownames(Cox_df),sce$PID)]
Cox_df$OS_months <- sce$OS_months[match(rownames(Cox_df),sce$PID)]

Cox_df$age <- sce$age[match(rownames(Cox_df),sce$PID)]
Cox_df$grade <- factor(sce$grade[match(rownames(Cox_df),sce$PID)], levels = c("3","2","1"))
Cox_df$pT_simple <- as.factor(sce$pT_simple[match(rownames(Cox_df),sce$PID)])
Cox_df$pN_simple <- as.factor(sce$pN_simple[match(rownames(Cox_df),sce$PID)])
Cox_df$pM_simple <- as.factor(sce$pM_simple[match(rownames(Cox_df),sce$PID)])

#Based on proportions
covariates <- Cox_df %>% select(-status_OS,-OS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple, -Immune_low) %>% colnames

multi_formula <- as.formula(paste("Surv(OS_months, status_OS) ~ age + grade + pT_simple + pN_simple + pM_simple +", paste(covariates, collapse = " + "),"+ Immune_low"))

multi_model <- coxph(multi_formula, data = Cox_df)

multi_res <- summary(multi_model)

plot_df_prop <- data.frame(group = rownames(multi_res$coefficients),
                      p_value = signif(multi_res$coefficients[,"Pr(>|z|)"],2),
                      HR = signif(multi_res$coefficients[,"exp(coef)"],2),
                      CI_low = signif(multi_res$conf.int[,"lower .95"],2),
                      CI_high = signif(multi_res$conf.int[,"upper .95"],2),
                      wald_test = signif(multi_res$waldtest["pvalue"],2))


#Plot hazard ratios
hr_prop <- ggplot(plot_df_prop, aes(x = reorder(group,dplyr::desc(HR)), y = HR, color = p_value < 0.05))+
geom_point(size = 4)+
geom_errorbar(aes(ymax = CI_high, ymin = CI_low))+
coord_cartesian(ylim=c(0,50))+
geom_hline(yintercept = 1,color = "black",linetype = "dashed")+
scale_color_brewer(palette = "Set1")+
xlab("Immune metaclusters")+
ylab("HR for con_prop")+
theme_classic()+
theme(axis.text.x = element_text(angle = 90))

hr_prop

## 2. Based on categories (mean division of cluster proportions)
Cox_df_mean <- foreach(i = Cox_df %>% select(-status_OS,-OS_months, -status_OS,-OS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames,.combine = "cbind")%do%{
Cox_df %>% select(i) %>% mutate(mean = ifelse(Cox_df[,i] > mean(Cox_df[,i]),"B","A")) %>% pull(mean)
}

colnames(Cox_df_mean) <- Cox_df %>% select(-status_OS,-OS_months, -status_OS,-OS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames
rownames(Cox_df_mean) <- Cox_df %>% select(-status_OS,-OS_months, -status_OS,-OS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% rownames
Cox_df_mean <- as.data.frame(Cox_df_mean)

identical(rownames(Cox_df), rownames(Cox_df_mean))

Cox_df_mean <- cbind(Cox_df_mean, Cox_df %>% select(status_OS,OS_months, status_OS, OS_months, age, grade, pT_simple, pN_simple, pM_simple))

#Based on proportions
covariates <- Cox_df_mean %>% select(-status_OS,-OS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames

multi_formula <- as.formula(paste("Surv(OS_months, status_OS) ~ age + grade + pT_simple + pN_simple + pM_simple +", paste(covariates, collapse = " + ")))

multi_model <- coxph(multi_formula, data = Cox_df_mean)

summary(multi_model)

multi_res <- summary(multi_model)

plot_df_mean <- data.frame(group = rownames(multi_res$coefficients),
                      p_value = signif(multi_res$coefficients[,"Pr(>|z|)"],2),
                      HR = signif(multi_res$coefficients[,"exp(coef)"],2),
                      CI_low = signif(multi_res$conf.int[,"lower .95"],2),
                      CI_high = signif(multi_res$conf.int[,"upper .95"],2),
                      wald_test = signif(multi_res$waldtest["pvalue"],2))

#Plot hazard ratios
hr_mean <- ggplot(plot_df_mean, aes(x = reorder(group,dplyr::desc(HR)), y = HR, color = p_value < 0.05))+
geom_point(size = 4)+
geom_errorbar(aes(ymax = CI_high, ymin = CI_low))+
coord_cartesian(ylim=c(0,50))+
geom_hline(yintercept = 1,color = "black",linetype = "dashed")+
scale_color_brewer(palette = "Set1")+
xlab("Immune metaclusters")+
ylab("HR for mean")+
theme_classic()+
theme(axis.text.x = element_text(angle = 90))

hr_mean

## 3. Based on categories (GMM division of cluster proportions)
Cox_df_GMM <- foreach(i = Cox_df %>% select(-status_OS,-OS_months, -status_OS,-OS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames,.combine = "cbind")%do%{
GMM_Cox <- Mclust(Cox_df[,i], G = 2)
as.character(GMM_Cox$classification)
}

colnames(Cox_df_GMM) <- Cox_df %>% select(-status_OS,-OS_months, -status_OS,-OS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames
rownames(Cox_df_GMM) <- Cox_df %>% select(-status_OS,-OS_months, -status_OS,-OS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% rownames
Cox_df_GMM <- as.data.frame(Cox_df_GMM,stringsAsFactors=TRUE)

identical(rownames(Cox_df), rownames(Cox_df_GMM))

Cox_df_GMM <- cbind(Cox_df_GMM, Cox_df %>% select(status_OS,OS_months, status_OS, OS_months, age, grade, pT_simple, pN_simple, pM_simple))

#Based on proportions
covariates <- Cox_df_GMM %>% select(-status_OS,-OS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames

multi_formula <- as.formula(paste("Surv(OS_months, status_OS) ~ age + grade + pT_simple + pN_simple + pM_simple +", paste(covariates, collapse = " + ")))

multi_model <- coxph(multi_formula, data = Cox_df_GMM)

summary(multi_model)

summary(Cox_df_GMM)

multi_res <- summary(multi_model)

plot_df_GMM <- data.frame(group = rownames(multi_res$coefficients),
                      p_value = signif(multi_res$coefficients[,"Pr(>|z|)"],2),
                      HR = signif(multi_res$coefficients[,"exp(coef)"],2),
                      CI_low = signif(multi_res$conf.int[,"lower .95"],2),
                      CI_high = signif(multi_res$conf.int[,"upper .95"],2),
                      wald_test = signif(multi_res$waldtest["pvalue"],2))

#Plot hazard ratios
hr_GMM <- ggplot(plot_df_GMM, aes(x = reorder(group,dplyr::desc(HR)), y = HR, color = p_value < 0.05))+
geom_point(size = 4)+
geom_errorbar(aes(ymax = CI_high, ymin = CI_low))+
coord_cartesian(ylim=c(0,50))+
geom_hline(yintercept = 1,color = "black",linetype = "dashed")+
scale_color_brewer(palette = "Set1")+
xlab("Immune metaclusters")+
ylab("HR for GMM")+
theme_classic()+
theme(axis.text.x = element_text(angle = 90))

hr_prop/
hr_mean/
hr_GMM
```
```{r OS survival_curves}
#Plot survival curves
fit <- survfit(Surv(OS_months, status_OS) ~ BnT_cell, data = Cox_df_mean)
print(fit)

p1 <- ggsurvplot(fit,
          pval = TRUE,
          risk.table = TRUE, # Add risk table
          risk.table.col = "strata", # Change risk table color by groups
          linetype = "strata", # Change line type by groups
          surv.median.line = "hv", # Specify median survival
          ggtheme = theme_classic(), # Change ggplot2 theme
          palette = "Set1",
          log.rank.weights = "n")+
  ggtitle("Mean proportion")

fit <- survfit(Surv(OS_months, status_OS) ~ Monocyte, data = Cox_df_mean)
print(fit)

p2 <- ggsurvplot(fit,
          pval = TRUE,
          risk.table = TRUE, # Add risk table
          risk.table.col = "strata", # Change risk table color by groups
          linetype = "strata", # Change line type by groups
          surv.median.line = "hv", # Specify median survival
          ggtheme = theme_classic(), # Change ggplot2 theme
          palette = "Set1",
          log.rank.weights = "n")+
  ggtitle("Mean proportion")

fit <- survfit(Surv(OS_months, status_OS) ~ CD8_T_cell, data = Cox_df_mean)
print(fit)

p2 <- ggsurvplot(fit,
          pval = TRUE,
          risk.table = TRUE, # Add risk table
          risk.table.col = "strata", # Change risk table color by groups
          linetype = "strata", # Change line type by groups
          surv.median.line = "hv", # Specify median survival
          ggtheme = theme_classic(), # Change ggplot2 theme
          palette = "Set1",
          log.rank.weights = "n")+
  ggtitle("Mean proportion")


arrange_ggsurvplots(list(p1,p2))
```

### For DFS on metaclusters

Multivariate 

```{r cox dfs}
library(survival)
library(survminer)

PID_metacluster <- colData(sce) %>% as.data.frame() %>% select(immune_metacluster,PID) %>% group_by(PID) %>% table() %>% as.data.frame
PID_metacluster <- PID_metacluster %>% group_by(PID) %>% mutate(fra = Freq/sum(Freq))

### Multivariate cox proportional hazard model on multiple covariates (here proportion of phenotypes)
PID_metagroups <- PID_metacluster %>% pivot_wider(id_cols = "PID",names_from = "immune_metacluster",values_from = "fra") %>% column_to_rownames("PID")

## 1. Based on continuous cluster proportions
Cox_df <- PID_metagroups

#add covariates for correction
Cox_df$status_DFS <- sce$status_DFS[match(rownames(Cox_df),sce$PID)]
Cox_df$DFS_months <- sce$DFS_months[match(rownames(Cox_df),sce$PID)]

Cox_df$age <- sce$age[match(rownames(Cox_df),sce$PID)]
Cox_df$grade <- factor(sce$grade[match(rownames(Cox_df),sce$PID)], levels = c("3","2","1"))
Cox_df$pT_simple <- as.factor(sce$pT_simple[match(rownames(Cox_df),sce$PID)])
Cox_df$pN_simple <- as.factor(sce$pN_simple[match(rownames(Cox_df),sce$PID)])
Cox_df$pM_simple <- as.factor(sce$pM_simple[match(rownames(Cox_df),sce$PID)])

#Based on proportions
covariates <- Cox_df %>% select(-status_DFS,-DFS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple, -Immune_low) %>% colnames

multi_formula <- as.formula(paste("Surv(DFS_months, status_DFS) ~ age + grade + pT_simple + pN_simple + pM_simple +", paste(covariates, collapse = " + "),"+ Immune_low"))

multi_model <- coxph(multi_formula, data = Cox_df)

multi_res <- summary(multi_model)

plot_df_prop <- data.frame(group = rownames(multi_res$coefficients),
                      p_value = signif(multi_res$coefficients[,"Pr(>|z|)"],2),
                      HR = signif(multi_res$coefficients[,"exp(coef)"],2),
                      CI_low = signif(multi_res$conf.int[,"lower .95"],2),
                      CI_high = signif(multi_res$conf.int[,"upper .95"],2),
                      wald_test = signif(multi_res$waldtest["pvalue"],2))

write.csv(plot_df_prop, "/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/supp_tables/Cox_Multi_TMEMC.csv")


#Plot hazard ratiDFS
hr_prop <- ggplot(plot_df_prop, aes(x = reorder(group,dplyr::desc(HR)), y = HR, color = p_value < 0.05))+
geom_point(size = 4)+
geom_errorbar(aes(ymax = CI_high, ymin = CI_low))+
coord_cartesian(ylim=c(0.001,100))+
geom_hline(yintercept = 1,color = "black",linetype = "dashed")+
scale_color_brewer(palette = "Set1")+
xlab("Immune metaclusters")+
ylab("HR for con_prop")+
scale_y_continuous(trans = "log10")+
theme_classic()+
theme(axis.text.x = element_text(angle = 90))

hr_prop

## 2. Based on categories (mean division of cluster proportions)

Cox_df_mean <- foreach(i = Cox_df %>% select(-status_DFS,-DFS_months, -status_DFS,-DFS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames,.combine = "cbind")%do%{
Cox_df %>% select(i) %>% mutate(mean = ifelse(Cox_df[,i] > mean(Cox_df[,i]),"B","A")) %>% pull(mean)
}

colnames(Cox_df_mean) <- Cox_df %>% select(-status_DFS,-DFS_months, -status_DFS,-DFS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames
rownames(Cox_df_mean) <- Cox_df %>% select(-status_DFS,-DFS_months, -status_DFS,-DFS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% rownames
Cox_df_mean <- as.data.frame(Cox_df_mean)

identical(rownames(Cox_df), rownames(Cox_df_mean))

Cox_df_mean <- cbind(Cox_df_mean, Cox_df %>% select(status_DFS,DFS_months, status_DFS, DFS_months, age, grade, pT_simple, pN_simple, pM_simple))

#Based on proportions
covariates <- Cox_df_mean %>% select(-status_DFS,-DFS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames

multi_formula <- as.formula(paste("Surv(DFS_months, status_DFS) ~ age + grade + pT_simple + pN_simple + pM_simple +", paste(covariates, collapse = " + ")))

multi_model <- coxph(multi_formula, data = Cox_df_mean)

summary(multi_model)

multi_res <- summary(multi_model)

plot_df_mean <- data.frame(group = rownames(multi_res$coefficients),
                      p_value = signif(multi_res$coefficients[,"Pr(>|z|)"],2),
                      HR = signif(multi_res$coefficients[,"exp(coef)"],2),
                      CI_low = signif(multi_res$conf.int[,"lower .95"],2),
                      CI_high = signif(multi_res$conf.int[,"upper .95"],2),
                      wald_test = signif(multi_res$waldtest["pvalue"],2))

  
#Plot hazard ratios
hr_mean <- ggplot(plot_df_mean, aes(x = reorder(group,dplyr::desc(HR)), y = HR, color = p_value < 0.05))+
geom_point(size = 4)+
geom_errorbar(aes(ymax = CI_high, ymin = CI_low))+
coord_cartesian(ylim=c(0,50))+
geom_hline(yintercept = 1,color = "black",linetype = "dashed")+
scale_color_brewer(palette = "Set1")+
xlab("Immune metaclusters")+
ylab("HR for mean")+
coord_cartesian(ylim=c(0.1,200))+
scale_y_continuous(trans = "log10")+
theme_classic()+
theme(axis.text.x = element_text(angle = 90))

hr_mean

## 3. Based on categories (GMM division of cluster proportions)
Cox_df_GMM <- foreach(i = Cox_df %>% select(-status_DFS,-DFS_months, -status_DFS,-DFS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames,.combine = "cbind")%do%{
GMM_Cox <- Mclust(Cox_df[,i], G = 2)
as.character(GMM_Cox$classification)
}

colnames(Cox_df_GMM) <- Cox_df %>% select(-status_DFS,-DFS_months, -status_DFS,-DFS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames
rownames(Cox_df_GMM) <- Cox_df %>% select(-status_DFS,-DFS_months, -status_DFS,-DFS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% rownames
Cox_df_GMM <- as.data.frame(Cox_df_GMM,stringsAsFactors=TRUE)

identical(rownames(Cox_df), rownames(Cox_df_GMM))

Cox_df_GMM <- cbind(Cox_df_GMM, Cox_df %>% select(status_DFS,DFS_months, status_DFS, DFS_months, age, grade, pT_simple, pN_simple, pM_simple))

#Based on proportions
covariates <- Cox_df_GMM %>% select(-status_DFS,-DFS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames

multi_formula <- as.formula(paste("Surv(DFS_months, status_DFS) ~ age + grade + pT_simple + pN_simple + pM_simple +", paste(covariates, collapse = " + ")))

multi_model <- coxph(multi_formula, data = Cox_df_GMM)

summary(multi_model)

summary(Cox_df_GMM)

multi_res <- summary(multi_model)

plot_df_GMM <- data.frame(group = rownames(multi_res$coefficients),
                      p_value = signif(multi_res$coefficients[,"Pr(>|z|)"],2),
                      HR = signif(multi_res$coefficients[,"exp(coef)"],2),
                      CI_low = signif(multi_res$conf.int[,"lower .95"],2),
                      CI_high = signif(multi_res$conf.int[,"upper .95"],2),
                      wald_test = signif(multi_res$waldtest["pvalue"],2))

#Plot hazard ratiDFS
hr_GMM <- ggplot(plot_df_GMM, aes(x = reorder(group,dplyr::desc(HR)), y = HR, color = p_value < 0.05))+
geom_point(size = 4)+
geom_errorbar(aes(ymax = CI_high, ymin = CI_low))+
coord_cartesian(ylim=c(0,50))+
geom_hline(yintercept = 1,color = "black",linetype = "dashed")+
scale_color_brewer(palette = "Set1")+
xlab("Immune metaclusters")+
ylab("HR for GMM")+
theme_classic()+
theme(axis.text.x = element_text(angle = 90))

hr_GMM

hr_prop/
hr_mean/
hr_GMM
```

Multivariate without M1 patients

```{r cox dfs}
library(survival)
library(survminer)

PID_metacluster <- colData(sce) %>% as.data.frame() %>% filter(pM_simple != 1) %>% select(immune_metacluster,PID) %>% group_by(PID) %>% table() %>% as.data.frame
PID_metacluster <- PID_metacluster %>% group_by(PID) %>% mutate(fra = Freq/sum(Freq))

### Multivariate cox proportional hazard model on multiple covariates (here proportion of phenotypes)
PID_metagroups <- PID_metacluster %>% pivot_wider(id_cols = "PID",names_from = "immune_metacluster",values_from = "fra") %>% column_to_rownames("PID")

## 1. Based on continuous cluster proportions
Cox_df <- PID_metagroups

#add covariates for correction
Cox_df$status_DFS <- sce$status_DFS[match(rownames(Cox_df),sce$PID)]
Cox_df$DFS_months <- sce$DFS_months[match(rownames(Cox_df),sce$PID)]

Cox_df$age <- sce$age[match(rownames(Cox_df),sce$PID)]
Cox_df$grade <- factor(sce$grade[match(rownames(Cox_df),sce$PID)], levels = c("3","2","1"))
Cox_df$pT_simple <- as.factor(sce$pT_simple[match(rownames(Cox_df),sce$PID)])
Cox_df$pN_simple <- as.factor(sce$pN_simple[match(rownames(Cox_df),sce$PID)])

#Based on proportions
covariates <- Cox_df %>% select(-status_DFS,-DFS_months, -age, -grade, -pT_simple, -pN_simple, -Immune_low) %>% colnames

multi_formula <- as.formula(paste("Surv(DFS_months, status_DFS) ~ age + grade + pT_simple + pN_simple +", paste(covariates, collapse = " + "),"+ Immune_low"))

multi_model <- coxph(multi_formula, data = Cox_df)

multi_res <- summary(multi_model)

plot_df_prop <- data.frame(group = rownames(multi_res$coefficients),
                      p_value = signif(multi_res$coefficients[,"Pr(>|z|)"],2),
                      HR = signif(multi_res$coefficients[,"exp(coef)"],2),
                      CI_low = signif(multi_res$conf.int[,"lower .95"],2),
                      CI_high = signif(multi_res$conf.int[,"upper .95"],2),
                      wald_test = signif(multi_res$waldtest["pvalue"],2))

#write.csv(plot_df_prop, "/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/supp_tables/Cox_Multi_TMEMC.csv")


#Plot hazard ratiDFS
hr_prop <- ggplot(plot_df_prop, aes(x = reorder(group,dplyr::desc(HR)), y = HR, color = p_value < 0.05))+
geom_point(size = 4)+
geom_errorbar(aes(ymax = CI_high, ymin = CI_low))+
coord_cartesian(ylim=c(0.0001,100000))+
geom_hline(yintercept = 1,color = "black",linetype = "dashed")+
scale_color_brewer(palette = "Set1")+
xlab("Immune metaclusters")+
ylab("HR for con_prop")+
scale_y_continuous(trans = "log10")+
theme_classic()+
theme(axis.text.x = element_text(angle = 90))

hr_prop
```

### Plot proportion density curves for all metaclusters

Add lines for mean and GMM division.

```{r}
#mean division
mean_metacluster <- PID_metacluster %>% group_by(immune_metacluster) %>% transmute(immune_metacluster,mean = mean(fra)) %>% unique()

#GMM division
GMM_div <- Cox_df_GMM %>% select(1:10) %>% rownames_to_column("PID") %>% pivot_longer(cols = 2:11,names_to = "immunemetacluster", values_to = "GMM")
PID_metacluster$GMM <- GMM_div$GMM
GMM_1_max_metacluster <- PID_metacluster %>% filter(GMM == "1") %>% group_by(immune_metacluster) %>% summarise(max = max(fra))

# Plot proportion density curves for all metaclusters
ggplot(PID_metacluster, aes(fra))+
  facet_wrap(~immune_metacluster)+
  geom_density(aes(fill = immune_metacluster))+
  geom_rug()+
  coord_cartesian(ylim = c(0,10))+
  scale_fill_manual(values = metadata(sce)$colors$immune_metacluster)+
  theme_classic()+
  geom_vline(data = mean_metacluster, aes(xintercept = mean),color = "#E41A1C", linetype = "dashed")+#red
  geom_vline(data = GMM_1_max_metacluster, aes(xintercept = max),color = "#377EB8", linetype = "dashed") #blue
```

### For OS on metacluster density

```{r}
library(survival)
library(survminer)

PID_metacluster <- colData(sce) %>% as.data.frame() %>% select(immune_metacluster,PID) %>% group_by(PID) %>% table() %>% as.data.frame

PID_area <- colData(sce) %>% as.data.frame() %>% select(tissue_area,PID) %>% unique() %>% group_by(PID) %>% mutate(mean_tissue_area = mean(tissue_area))

PID_metacluster$mean_tissue_area <- PID_area$mean_tissue_area[match(PID_metacluster$PID, PID_area$PID)]
  
PID_metacluster <- PID_metacluster %>% group_by(PID) %>% mutate(density = Freq/(mean_tissue_area/1000000))  

### Multivariate cox proportional hazard model on multiple covariates (here density of phenotypes)
PID_metagroups <- PID_metacluster %>% pivot_wider(id_cols = "PID",names_from = "immune_metacluster",values_from = "density") %>% column_to_rownames("PID")

#colnames(PID_metagroups) <- paste0("MC",colnames(PID_metagroups))

## 1. Based on continuous cluster proportions
Cox_df <- PID_metagroups

#add covariates for correction
Cox_df$status_OS <- sce$status_OS[match(rownames(Cox_df),sce$PID)]
Cox_df$OS_months <- sce$OS_months[match(rownames(Cox_df),sce$PID)]

Cox_df$age <- sce$age[match(rownames(Cox_df),sce$PID)]
Cox_df$grade <- factor(sce$grade[match(rownames(Cox_df),sce$PID)], levels = c("3","2","1"))
Cox_df$pT_simple <- as.factor(sce$pT_simple[match(rownames(Cox_df),sce$PID)])
Cox_df$pN_simple <- as.factor(sce$pN_simple[match(rownames(Cox_df),sce$PID)])
Cox_df$pM_simple <- as.factor(sce$pM_simple[match(rownames(Cox_df),sce$PID)])

#mean division 
Cox_df_mean <- foreach(i = Cox_df %>% select(-status_OS,-OS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames,.combine = "cbind")%do%{
Cox_df %>% select(i) %>% mutate(mean = ifelse(Cox_df[,i] > mean(Cox_df[,i]),"B","A")) %>% pull(mean)
}

colnames(Cox_df_mean) <- Cox_df %>% select(-status_OS,-OS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames
rownames(Cox_df_mean) <- Cox_df %>% select(-status_OS,-OS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% rownames
Cox_df_mean <- as.data.frame(Cox_df_mean)

identical(rownames(Cox_df), rownames(Cox_df_mean))

Cox_df_mean <- cbind(Cox_df_mean, Cox_df %>% select(status_OS,OS_months, age, grade, pT_simple, pN_simple, pM_simple))

#run the model
covariates <- Cox_df_mean %>% select(-status_OS,-OS_months, -age, -grade, -pT_simple, -pN_simple, -pM_simple) %>% colnames

multi_formula <- as.formula(paste("Surv(OS_months, status_OS) ~ age + grade + pT_simple + pN_simple + pM_simple +", paste(covariates, collapse = " + ")))

multi_model <- coxph(multi_formula, data = Cox_df_mean)

multi_res <- summary(multi_model)

plot_df_prop <- data.frame(group = rownames(multi_res$coefficients),
                      p_value = signif(multi_res$coefficients[,"Pr(>|z|)"],2),
                      HR = signif(multi_res$coefficients[,"exp(coef)"],2),
                      CI_low = signif(multi_res$conf.int[,"lower .95"],2),
                      CI_high = signif(multi_res$conf.int[,"upper .95"],2),
                      wald_test = signif(multi_res$waldtest["pvalue"],2))


#Plot hazard ratios
hr_prop <- ggplot(plot_df_prop, aes(x = reorder(group,desc(HR)), y = HR, color = p_value < 0.05))+
geom_point(size = 4)+
geom_errorbar(aes(ymax = CI_high, ymin = CI_low))+
coord_cartesian(ylim=c(0,50))+
geom_hline(yintercept = 1,color = "black",linetype = "dashed")+
scale_color_brewer(palette = "Set1")+
xlab("Immune metaclusters")+
ylab("HR for density")+
theme_classic()+
theme(axis.text.x = element_text(angle = 90))

hr_prop
```

## Save sce objects

```{r save_objects, message = FALSE, warning=FALSE}
#Save sce
saveRDS(sce, "/mnt/central_nas/tnbc_volume/TNBC/data_analysis/06_sce_TNBC_TME_new.rds")
```

## Save plots as pdf

```{r save_plots}
##IMPORTANT: change to chunk output in console for this last part 

##save plots as pdf (vector graphics)
pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_Heatmap_ClusterTesting.pdf",height = 4, width = 9)
p1+p2+p3 
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_Silhouette_Cluster130.pdf",width = 12,height = 10)
p4
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_HeatmapCluster.pdf",width = 10,height = 6)
draw(h,annotation_legend_list = list(lgd_astir, lgd_meta))
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_HeatmapCluster_Scales.pdf",width = 13,height = 8)
h_exprs + h_min_max
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_ClusterAnno.pdf",width = 5,height = 8)
h_cluster_anno
dev.off()


pdf("/Volumes/rcc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_HeatmapCluster_subcells.pdf",width = 10,height = 8)
dittoHeatmap(sce_sub, 
             genes = rownames(sce_sub)[rowData(sce_sub)$cluster_channel],
             assay = "min_max", 
             scale = "none", 
             order.by = "cluster_50", 
             heatmap.colors = viridis(50), 
             annot.by = c("cluster_50"),
             annot.colors=col_clusters$cluster_50[order(as.numeric(names(col_clusters$cluster_50)))])
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_ExpressionCor.pdf",width = 10,height = 8)
corrplot(cor(cor_sce,method = "spearman"), method="color", col=brewer.pal(n=8, name="RdBu"),  
         type="upper", order="hclust",hclust.method = "ward.D2", 
         addCoef.col = "black", addCoefasPercent = F,# Add coefficient of correlation
         tl.col="black", tl.srt=40, #Text label color and rotation
         # Combine with significance
         #p.mat = p.mat, sig.level = 0.01, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag=F
)
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_TSNECluster.pdf",width = 8,height = 6)
TSNE_clus
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_TSNEExpression.pdf",width = 12,height = 6)
exp_plots_TSNE
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_UMAPExpression.pdf",width = 12,height = 6)
exp_plots_UMAP
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_ClusterPIDSID.pdf",width = 12,height = 6)
p_clusterPID
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_PIDCluster_ClusterCor.pdf",width = 14,height = 6)
(d_PID + SID_immune_metacluster_cor_plot)/
(d_PID_dens + SID_metacluster_dens_cor_plot)
dev.off()

pdf("/Volumes/rcc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_HeatmapPIDgroups.pdf",width = 11,height = 7)
h_g
dev.off()

pdf("/Volumes/rcc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_HeatmapPIDgroupsCLR.pdf",width = 11,height = 7)
h_g_1
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_CC_Cluster_HR_DFS.pdf",width = 8,height = 10)
hr_prop/
hr_cat/
hr_cat_GMM
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_MC_HR_Multi_DFS.pdf",width = 6,height = 3, onefile = FALSE)
hr_prop
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_FractionHeatmap.pdf", width = 8, height = 18, onefile = FALSE)
fraction_anno + NULL
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_Metacluster_Fra_Cor.pdf", width = 6)
SID_immune_metacluster_cor_plot
dev.off()


pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_clrDR_MetaclusterPID.pdf", height = 5, width = 6)
clrDR(sce_cat, dr = "PCA", 
      by = "cluster_id", k = "immune_metacluster", 
      label_by = "cluster_id", arrow_col = "PID", 
      point_pal = metadata(sce_cat)$colors$immune_metacluster, 
      arrows = FALSE) +
  scale_color_manual(values = metadata(sce_cat)$colors$PID)+
  guides(color = "none")+
  theme_classic()
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_clrDR_PIDMetacluster.pdf", height = 5, width = 6)
clrDR(sce_cat, dr = "PCA", 
      by = "sample_id", point_col = "sample_id",
      k = "immune_metacluster", point_pal = metadata(sce_cat)$colors$PID)+
  scale_color_manual(values = metadata(sce_cat)$colors$immune_metacluster)+
  guides(fill = "none")+
  theme_classic()
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_Cox_HR_DFS_Meta_Multi.pdf", height = 4, width = 8)
hr_prop
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_Cox_HR_DFS_Meta_Mean.pdf", height = 3, width = 6)
hr_mean
dev.off()

pdf("/mnt/central_nas/tnbc_volume/TNBC/data_analysis/01_figures/06_TME/TNBC_TME_MC_HR_Multi_DFS_M1out.pdf", height = 3, width = 6)
hr_prop
dev.off()

##IMPORTANT: change back to chunk output inline
```
